requirements.txt
python-telegram-bot
sqlalchemy
chromadb==0.4.18
apscheduler
PyPDF2==3.0.1
aiohttp==3.9.1
pyyaml==6.0.1
beautifulsoup4==4.12.2
google-auth
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
requests==2.31.0
python-dateutil==2.8.2
sentence-transformers==2.2.2
lxml==4.9.3
rich
python-dotenv==1.0.0
pytz==2023.3
dateparser==1.2.0
fpdf==1.7.2
youtube-transcript-api==0.6.1
feedparser==6.0.10
